`JournalBuilder` -> `QueueStop`
